+++ src/components/productItem/ProductItem.jsx
@@ -10,7 +10,7 @@ import Text from 'components/base/text/Text';
 import { ReactComponent as PlusIcon } from 'assets/plus_icon.svg';
 import { ReactComponent as MinusIcon } from 'assets/minus_icon.svg';
 
-import { postShoppingCartItem, putShoppingCartItem } from 'modules/shoppingCarts';
+import { postShoppingCartItem, putShoppingCartItem } from 'middlewares/shoppingCarts';
 
 import {
   StyledProductItem,
